move injector methods from MinigameService to GuiceManager
