+++ src/main/java/io/github/minigamecore/api/util/manager/package-info.java
@@ -23,38 +23,7 @@
  * THE SOFTWARE.
  */
 
-package io.github.minigamecore.api;
+@ParametersAreNonnullByDefault
+package io.github.minigamecore.api.util.manager;
 
-import io.github.minigamecore.api.util.config.ConfigurationManager;
-import io.github.minigamecore.api.util.manager.GuiceManager;
-import org.spongepowered.api.service.ServiceManager;
-
-import javax.annotation.Nonnull;
-
-/**
- * The main service provided by MinigameCore.
- *
- * <p>
- *     This service is guaranteed to be present if the MinigameCore plugin is active. Therefore {@link ServiceManager#provideUnchecked(Class)} can be
- *     safely used. {@link MinigameService} can also be injected.
- * </p>
- */
-public interface MinigameService {
-
-    /**
-     * Gets the {@link ConfigurationManager}.
-     *
-     * @return The ConfigurationManager.
-     */
-    @Nonnull
-    ConfigurationManager getConfigurationManager();
-
-    /**
-     * Gets the {@link ConfigurationManager}.
-     *
-     * @return The ConfigurationManager.
-     */
-    @Nonnull
-    GuiceManager getGuiceManager();
-
-}
+import javax.annotation.ParametersAreNonnullByDefault;
